egui_glow: move vao code to own file
