+++ egui_glow/src/lib.rs
@@ -14,7 +14,7 @@ pub use painter::Painter;
 mod misc_util;
 mod post_process;
 mod shader_version;
-mod vao_emulate;
+mod vao;
 
 #[cfg(all(not(target_arch = "wasm32"), feature = "winit"))]
 pub mod winit;
